added flex elements to coulums for score board
in the html split scoreboard into a flex column for home away and time and moved inmages and team name into seperate rows.
in the css also added  flex-basis: to home and away team to have a fixed length of 100 pixels to fix an lignment issue betewwen team names and photos.